Rewrite gDM type inference for better extraction of `dispatch`
`getDefaultMiddleware` defines an initial type that may or may not
include `thunk`, depending on options. From there, `EnhancedStore`
tries to extract any extensions to `dispatch` from "the type of all
middleware as an array", to fully infer the correct type of
`store.dispatch`.

This was flaky. In particular, the `MiddlewareArray` type ended up
with "a union of all middleware" type, and that meant the extensions
weren't always getting read correctly.

This commit rewrites the inference to be more correct:

- `MiddlewareArray` now tracks an exact tuple type for its contents
- `.concat/prepend` now update that middleware tuple type
- We exactly extract the dispatch extensions from the tuple,
  and retain the exact order of the middleware as declared
- All the extensions are intersected together with `Dispatch`

This should now correctly represent cases like the listener
middleware returning an `Unsubscribe` callback in response to
`dispatch(addListenerAction())`

This commit also drops support for passing `null` as the `extraArg`
generic for `ThunkAction`, as that is an outdated pattern and
we want to discourage it.